requirements.txt
Flask>=2.0
Flask-SocketIO>=5.0
google-generativeai>=0.5.0
python-dotenv
eventlet